+++ extension/src/popup/components/account/NotFundedMessage/styles.scss
@@ -1,0 +1,31 @@
+.NotFunded {
+  background: var(--pal-background-secondary);
+  border-radius: 0.5rem;
+  margin-top: 2rem;
+  padding: 1rem;
+
+  &__header {
+    display: flex;
+    align-items: center;
+    color: var(--pal-text-primary);
+    font-size: 0.875rem;
+    line-height: 1.5rem;
+    font-weight: var(--font-weight-medium);
+    text-transform: uppercase;
+    padding-bottom: 0.25rem;
+  }
+
+  &__copy {
+    font-size: 0.875rem;
+    line-height: 1.375rem;
+    padding-top: 0.5rem;
+  }
+
+  &__links {
+    padding-top: 0.5rem;
+  }
+
+  &__form {
+    padding-top: 3.375rem;
+  }
+}
